rename, origin.py and coordinate_systems.py
fix name conflicts, remove hard code in origin.py, and revert to the original types in coordinate_systems.py
