+++ a5/core/coordinate_systems.py
@@ -4,16 +4,17 @@ SPDX-License-Identifier: Apache-2.0
 Copyright (c) A5 contributors
 """
 
+from typing import NewType, Tuple, TypeVar, Union
 import numpy as np
-from typing import NewType, Tuple, TypeVar, Union, Any
+from numpy.typing import NDArray
 
 # Base types
 Degrees = NewType('Degrees', float)
 Radians = NewType('Radians', float)
 
 # Vector types
-Vec2 = np.ndarray[Any, np.dtype[np.float64]]  # 2D vector
-Vec3 = np.ndarray[Any, np.dtype[np.float64]]  # 3D vector
+Vec2 = NDArray[np.float64]  # 2D vector
+Vec3 = NDArray[np.float64]  # 3D vector
 
 # 2D coordinate systems
 """
