+++ a5/core/coordinate_transforms.py
@@ -7,7 +7,7 @@ Copyright (c) A5 contributors
 import math
 import numpy as np
 from typing import cast
-from .a5_types import (
+from .coordinate_systems import (
     Degrees, Radians, Face, Polar, IJ, Cartesian, Spherical, LonLat,
     Vec2, Vec3
 )
